bugfix: Nominatives that are considered names are now processed
I had a problem with nouns that were assigned as a Name by the
RFTagger. The nouns are now recognized and analyzed. I had
to incooperate genitives in order to make the whole thing work.
The concept map might blow up as a consequence of this design
decision. We'll see.